requirements.txt
Flask==2.2.5
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.5.0
opencv-python==4.8.0
gunicorn==20.1.0
h5py==3.8.0
